+++ packages/gui/src/bridge/auto-start/backend.js
@@ -20,7 +20,7 @@ EOF
 `
     await DevSidecar.api.shell.exec(cmd)
   } else {
-    const removeStart = 'sudo rm ~/.config/autostart/dev-sidecar.desktop -rf'
+    const removeStart = 'rm ~/.config/autostart/dev-sidecar.desktop -rf'
     await DevSidecar.api.shell.exec(removeStart)
   }
 }